bump

+++ Shared/Extension/Int.swift
@@ -1,5 +1,5 @@
 //
-//  Int64.swift
+//  Int.swift
 //  Netliphy
 //
 //  Created by Дмитрий Лисин on 02.04.2021.
@@ -7,11 +7,11 @@
 
 import Foundation
 
-extension Int64 {
+extension Int {
     var byteSize: String {
         let bcf = ByteCountFormatter()
         bcf.allowedUnits = .useAll
         bcf.countStyle = .file
-        return bcf.string(fromByteCount: self)
+        return bcf.string(fromByteCount: Int64(self))
     }
 }
